design: PartSection 반응형 및 section 공통 style 분리
